permissions

+++ backend/src/resolvers/Mutation.js
@@ -7,10 +7,10 @@ const { checkIfLoggedIn, hasPermission } = require('../utils');
 
 const setCookie = (userId, ctx) => {
   const token = jwt.sign({ userId }, process.env.APP_SECRET);
-    ctx.response.cookie('token', token, {
-      httpOnly: true, // stop js access
-      maxAge: 1000 * 60 * 60 * 24 * 365 // 1 year 
-    });
+  ctx.response.cookie('token', token, {
+    httpOnly: true, // stop js access
+    maxAge: 1000 * 60 * 60 * 24 * 365 // 1 year 
+  });
 }
 
 const Mutation = {
@@ -43,10 +43,17 @@ const Mutation = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // find the items
-    const item =  await ctx.db.query.item({ where }, `{ id title}`);
+    const item = await ctx.db.query.item({ where }, `{ id title, user { id }}`);
+    const { userId, user } = ctx.request;
+    const ownsItem = item.user.id === userId;
+    const hasPermission = user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+
     // check if they own the item
-    // delete it
+    if (!ownsItem && !hasPermission) {
+      throw new Error(`You don't have permission to delete`);
+    }
     return ctx.db.mutation.deleteItem({ where }, info);
+    // delete it
   },
   async signup(parent, args, ctx, info) {
     args.email = args.email.toLowerCase();
@@ -66,7 +73,7 @@ const Mutation = {
   },
   async signin(parent, { email, password }, ctx, info) {
     // check for user with email
-    const user = await ctx.db.query.user({ where: { email }});
+    const user = await ctx.db.query.user({ where: { email } });
     if (!user) {
       throw new Error(`No such user found for email ${email}`);
     }
@@ -78,13 +85,13 @@ const Mutation = {
     setCookie(user.id, ctx);
     return user;
   },
-  async signout(parent, args, ctx, info){
+  async signout(parent, args, ctx, info) {
     ctx.response.clearCookie('token')
     return { message: 'Goodbye' };
   },
-  async requestReset(parent, { email }, ctx, info){
+  async requestReset(parent, { email }, ctx, info) {
     // check if real user
-    const user = await ctx.db.query.user({ where: { email }});
+    const user = await ctx.db.query.user({ where: { email } });
     if (!user) {
       throw new Error(`No such user found for email ${email}`);
     }
@@ -96,7 +103,7 @@ const Mutation = {
       where: { email },
       data: { resetToken, resetTokenExpiry }
     });
-    
+
     const mailResponse = await transport.sendMail({
       from: 'tomascollins@gmail.com',
       to: email,
@@ -104,7 +111,7 @@ const Mutation = {
       html: makeEmail(`Your password reset token is here! \n\n
       <a href="${process.env.FRONTEND_URL}/reset?resetToken=${resetToken}">Click here to reset</a>`)
     });
-    
+
     return { message: 'Thanks' };
   },
   async resetPassword(parent, args, ctx, info) {
@@ -130,22 +137,59 @@ const Mutation = {
         resetTokenExpiry: null
       }
     });
-    setCookie(user.userId, ctx); 
+    setCookie(user.userId, ctx);
     return updatedUser;
   },
   async updatePermissions(parent, args, ctx, info) {
     checkIfLoggedIn(ctx);
-    
-    const currentUser = ctx.db.query.user({
-      where: {
-        id: ctx.request.userId
-      }, info
-    });
+    const currentUser = await ctx.db.query.user(
+      { where: { id: ctx.request.userId } },
+      info
+    );
+
     hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
 
     return ctx.db.mutation.updateUser({
-      
+      data: {
+        permissions: {
+          set: args.permissions // needed for enum type
+        }
+      },
+      where: {
+        id: args.userId
+      },
+      info // query data for front end
     })
+  },
+  async addToCart(parent, args, ctx, info) {
+    // check if signed in
+    checkIfLoggedIn(ctx);
+    const { userId } = ctx.request;
+    // query the current cart
+    const [existingCartItem] = await ctx.db.query.cartItems({
+      where: {
+        user: { id: userId },
+        item: { id: args.id }
+      }
+    });
+    // check if that item is already in their cart
+    if (existingCartItem) {
+      console.log('already in cart');
+      return ctx.db.mutation.updateCartItem({
+        where: { id: existingCartItem.id },
+        data: { quantity: existingCartItem.quantity + 1 }
+      }, info);
+    }
+    return ctx.db.mutation.createCartItem({
+      data: {
+        user: {
+          connect: { id: userId }
+        },
+        item: {
+          connect: { id: args.id }
+        }
+      }
+    }, info)
   }
 };
